apps/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-dotenv==1.0.0
neo4j==5.14.1
aiofiles==23.2.1
httpx==0.27.2
prometheus-client==0.20.0
